+++ lib/models/src/order/order.freezed.dart
@@ -28,7 +28,7 @@ mixin _$Order {
   String get name => throw _privateConstructorUsedError;
   int get orderNumber => throw _privateConstructorUsedError;
   String get processedAt => throw _privateConstructorUsedError;
-  ShippingAddress get shippingAddress => throw _privateConstructorUsedError;
+  ShippingAddress? get shippingAddress => throw _privateConstructorUsedError;
   ShippingAddress? get billingAddress => throw _privateConstructorUsedError;
   String get statusUrl => throw _privateConstructorUsedError;
   PriceV2 get subtotalPriceV2 => throw _privateConstructorUsedError;
@@ -62,7 +62,7 @@ abstract class $OrderCopyWith<$Res> {
       String name,
       int orderNumber,
       String processedAt,
-      ShippingAddress shippingAddress,
+      ShippingAddress? shippingAddress,
       ShippingAddress? billingAddress,
       String statusUrl,
       PriceV2 subtotalPriceV2,
@@ -77,7 +77,7 @@ abstract class $OrderCopyWith<$Res> {
       List<SuccessfulFullfilment>? successfulFulfillments});
 
   $LineItemsOrderCopyWith<$Res> get lineItems;
-  $ShippingAddressCopyWith<$Res> get shippingAddress;
+  $ShippingAddressCopyWith<$Res>? get shippingAddress;
   $ShippingAddressCopyWith<$Res>? get billingAddress;
   $PriceV2CopyWith<$Res> get subtotalPriceV2;
   $PriceV2CopyWith<$Res> get totalPriceV2;
@@ -107,7 +107,7 @@ class _$OrderCopyWithImpl<$Res, $Val extends Order>
     Object? name = null,
     Object? orderNumber = null,
     Object? processedAt = null,
-    Object? shippingAddress = null,
+    Object? shippingAddress = freezed,
     Object? billingAddress = freezed,
     Object? statusUrl = null,
     Object? subtotalPriceV2 = null,
@@ -154,10 +154,10 @@ class _$OrderCopyWithImpl<$Res, $Val extends Order>
           ? _value.processedAt
           : processedAt // ignore: cast_nullable_to_non_nullable
               as String,
-      shippingAddress: null == shippingAddress
+      shippingAddress: freezed == shippingAddress
           ? _value.shippingAddress
           : shippingAddress // ignore: cast_nullable_to_non_nullable
-              as ShippingAddress,
+              as ShippingAddress?,
       billingAddress: freezed == billingAddress
           ? _value.billingAddress
           : billingAddress // ignore: cast_nullable_to_non_nullable
@@ -219,8 +219,12 @@ class _$OrderCopyWithImpl<$Res, $Val extends Order>
 
   @override
   @pragma('vm:prefer-inline')
-  $ShippingAddressCopyWith<$Res> get shippingAddress {
-    return $ShippingAddressCopyWith<$Res>(_value.shippingAddress, (value) {
+  $ShippingAddressCopyWith<$Res>? get shippingAddress {
+    if (_value.shippingAddress == null) {
+      return null;
+    }
+
+    return $ShippingAddressCopyWith<$Res>(_value.shippingAddress!, (value) {
       return _then(_value.copyWith(shippingAddress: value) as $Val);
     });
   }
@@ -298,7 +302,7 @@ abstract class _$$OrderImplCopyWith<$Res> implements $OrderCopyWith<$Res> {
       String name,
       int orderNumber,
       String processedAt,
-      ShippingAddress shippingAddress,
+      ShippingAddress? shippingAddress,
       ShippingAddress? billingAddress,
       String statusUrl,
       PriceV2 subtotalPriceV2,
@@ -315,7 +319,7 @@ abstract class _$$OrderImplCopyWith<$Res> implements $OrderCopyWith<$Res> {
   @override
   $LineItemsOrderCopyWith<$Res> get lineItems;
   @override
-  $ShippingAddressCopyWith<$Res> get shippingAddress;
+  $ShippingAddressCopyWith<$Res>? get shippingAddress;
   @override
   $ShippingAddressCopyWith<$Res>? get billingAddress;
   @override
@@ -349,7 +353,7 @@ class __$$OrderImplCopyWithImpl<$Res>
     Object? name = null,
     Object? orderNumber = null,
     Object? processedAt = null,
-    Object? shippingAddress = null,
+    Object? shippingAddress = freezed,
     Object? billingAddress = freezed,
     Object? statusUrl = null,
     Object? subtotalPriceV2 = null,
@@ -396,10 +400,10 @@ class __$$OrderImplCopyWithImpl<$Res>
           ? _value.processedAt
           : processedAt // ignore: cast_nullable_to_non_nullable
               as String,
-      shippingAddress: null == shippingAddress
+      shippingAddress: freezed == shippingAddress
           ? _value.shippingAddress
           : shippingAddress // ignore: cast_nullable_to_non_nullable
-              as ShippingAddress,
+              as ShippingAddress?,
       billingAddress: freezed == billingAddress
           ? _value.billingAddress
           : billingAddress // ignore: cast_nullable_to_non_nullable
@@ -500,7 +504,7 @@ class _$OrderImpl extends _Order {
   @override
   final String processedAt;
   @override
-  final ShippingAddress shippingAddress;
+  final ShippingAddress? shippingAddress;
   @override
   final ShippingAddress? billingAddress;
   @override
@@ -634,7 +638,7 @@ abstract class _Order extends Order {
       required final String name,
       required final int orderNumber,
       required final String processedAt,
-      required final ShippingAddress shippingAddress,
+      required final ShippingAddress? shippingAddress,
       required final ShippingAddress? billingAddress,
       required final String statusUrl,
       required final PriceV2 subtotalPriceV2,
@@ -668,7 +672,7 @@ abstract class _Order extends Order {
   @override
   String get processedAt;
   @override
-  ShippingAddress get shippingAddress;
+  ShippingAddress? get shippingAddress;
   @override
   ShippingAddress? get billingAddress;
   @override