+++ lib/models/src/order/order.g.dart
@@ -16,8 +16,10 @@ _$OrderImpl _$$OrderImplFromJson(Map<String, dynamic> json) => _$OrderImpl(
       name: json['name'] as String,
       orderNumber: json['orderNumber'] as int,
       processedAt: json['processedAt'] as String,
-      shippingAddress: ShippingAddress.fromJson(
-          json['shippingAddress'] as Map<String, dynamic>),
+      shippingAddress: json['shippingAddress'] == null
+          ? null
+          : ShippingAddress.fromJson(
+              json['shippingAddress'] as Map<String, dynamic>),
       billingAddress: json['billingAddress'] == null
           ? null
           : ShippingAddress.fromJson(