+++ lib/models/src/order/shipping_address/shipping_address.dart
@@ -7,12 +7,12 @@ part 'shipping_address.g.dart';
 class ShippingAddress with _$ShippingAddress {
   factory ShippingAddress({
     required String name,
-    required String firstName,
     required String id,
     required String lastName,
     required String address1,
     required String city,
     required String country,
+    String? firstName,
     String? address2,
     String? company,
     String? countryCodeV2,