+++ lib/models/src/order/shipping_address/shipping_address.g.dart
@@ -10,12 +10,12 @@ _$ShippingAddressImpl _$$ShippingAddressImplFromJson(
         Map<String, dynamic> json) =>
     _$ShippingAddressImpl(
       name: json['name'] as String,
-      firstName: json['firstName'] as String,
       id: json['id'] as String,
       lastName: json['lastName'] as String,
       address1: json['address1'] as String,
       city: json['city'] as String,
       country: json['country'] as String,
+      firstName: json['firstName'] as String?,
       address2: json['address2'] as String?,
       company: json['company'] as String?,
       countryCodeV2: json['countryCodeV2'] as String?,
@@ -31,12 +31,12 @@ Map<String, dynamic> _$$ShippingAddressImplToJson(
         _$ShippingAddressImpl instance) =>
     <String, dynamic>{
       'name': instance.name,
-      'firstName': instance.firstName,
       'id': instance.id,
       'lastName': instance.lastName,
       'address1': instance.address1,
       'city': instance.city,
       'country': instance.country,
+      'firstName': instance.firstName,
       'address2': instance.address2,
       'company': instance.company,
       'countryCodeV2': instance.countryCodeV2,